some changes

+++ core/management/commands/resetdb.py
@@ -12,12 +12,20 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         dbname = settings.DATABASES["default"]["NAME"]
-
         cursor = connection.cursor()
-        cursor.execute('show tables;')
-        parts = ('DROP TABLE IF EXISTS %s;' % table for (table,) in cursor.fetchall())
-        sql = 'SET FOREIGN_KEY_CHECKS = 0;\n' + '\n'.join(parts) + 'SET FOREIGN_KEY_CHECKS = 1;\n'
-        connection.cursor().execute(sql)
+        if 'db.sqlite3' in dbname:
+            pass
+            # try:
+            #     f = open("delete_data.sql", "r", encoding='UTF-8')
+            #     sql = f.read()
+            #     cursor.executescript(sql)
+            # except:
+            #     print("not_Deletd")
+        else:
+            cursor.execute('show tables;')
+            parts = ('DROP TABLE IF EXISTS %s;' % table for (table,) in cursor.fetchall())
+            sql = 'SET FOREIGN_KEY_CHECKS = 0;\n' + '\n'.join(parts) + 'SET FOREIGN_KEY_CHECKS = 1;\n'
+            connection.cursor().execute(sql)
         base = str(settings.BASE_DIR)
         migrations = glob.glob(os.path.join(base, "*", "migrations"))
         for migration in migrations:
@@ -26,3 +34,24 @@ class Command(BaseCommand):
         for app in apps:
             os.system("python manage.py makemigrations %s" % app)
         os.system("python manage.py migrate")
+        # if 'db.sqlite3' in dbname:
+        f = open("test_data.sql", "r", encoding='UTF-8')
+        sql = f.read()
+        cursor.executescript(sql)
+        # cursor.executescript(sql)
+
+    # def delete_all_tables(self, cursor):
+    #     tables = self.get_tables(cursor)
+    #     print(tables)
+    #     self.delete_tables(cursor, tables)
+    #
+    # def get_tables(self, cursor):
+    #     cursor.execute("SELECT name FROM sqlite_schema WHERE type='table';")
+    #     tables = cursor.fetchall()
+    #     return tables
+    #
+    # def delete_tables(self, cursor, tables):
+    #     TABLE_PARAMETER = "{TABLE_PARAMETER}"
+    #     for table, in tables:
+    #         sql = f"DROP TABLE {TABLE_PARAMETER};".replace("SELECT name FROM sqlite_schema WHERE type='table';", table)
+    #         cursor.execute(sql)
